// src/app/Model/EtiquetaRequestForm.ts
export interface EtiquetaRequestForm{
    direccionEntrega: string;
    destinoFinal: string
    proveedor: string
    rfc: string
    dependencia: string
    clave: string
    denominacionGenerica: string;
    lote: string;
    cantidadDeEtiquetas: number;
    formaFarmaceutica: string;
    descripcion: string;
    fabricante: string;
    regSanitario: string;
    fechaFabricacion: string;
    fechaCaducidad: string;
    remision: string;
    numeroContrato: string;
    numeroOrden: string;
    numeroLicitacion: string;
    tipoEtiqueta: string
    codigoBarras: string
    lotQuantity: string
    cantidadCustom: number
    nombreentrega: string
    nombrefinal: string
    nombreprov: string
    idImpresora: number
}